Moved ColorAdapter to its own file, removed fields
