+++ app/src/main/java/com/amaze/filemanager/adapters/ColorAdapter.java
@@ -1,0 +1,87 @@
+package com.amaze.filemanager.adapters;
+
+import android.content.Context;
+import android.graphics.drawable.GradientDrawable;
+import android.support.annotation.ColorInt;
+import android.support.annotation.ColorRes;
+import android.support.annotation.NonNull;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+import android.widget.AdapterView;
+import android.widget.ArrayAdapter;
+import android.widget.ImageView;
+
+import com.amaze.filemanager.R;
+import com.amaze.filemanager.utils.Utils;
+import com.amaze.filemanager.utils.color.ColorUsage;
+
+import java.util.List;
+
+public class ColorAdapter extends ArrayAdapter<Integer> implements AdapterView.OnItemClickListener {
+    private ColorUsage usage;
+    @ColorInt
+    private int selectedColor;
+    private OnColorSelected onColorSelected;
+
+    /**
+     * Constructor for adapter that handles the view creation of color chooser dialog in preferences
+     *
+     * @param context the context
+     * @param colors  array list of color hex values in form of string; for the views
+     * @param usage   the preference usage for setting new selected color preference value
+     * @param selectedColor currently selected color
+     * @param l OnColorSelected listener for when a color is selected
+     */
+    public ColorAdapter(Context context, List<Integer> colors, ColorUsage usage,
+                        @ColorInt int selectedColor, OnColorSelected l) {
+        super(context, R.layout.rowlayout, colors);
+        this.usage = usage;
+        this.selectedColor = selectedColor;
+        this.onColorSelected = l;
+    }
+
+    @ColorInt
+    private int getColor(@ColorRes int colorRes) {
+        return Utils.getColor(getContext(), colorRes);
+    }
+
+    @ColorRes
+    private int getColorResAt(int position) {
+        Integer item = getItem(position);
+
+        if (item == null) {
+            return usage.getDefaultColor();
+        } else {
+            return item;
+        }
+    }
+
+    @NonNull
+    @Override
+    public View getView(final int position, View convertView, @NonNull ViewGroup parent) {
+        LayoutInflater inflater = (LayoutInflater) getContext().getSystemService(Context.LAYOUT_INFLATER_SERVICE);
+        //TODO solve unconditional layout inflation
+        View rowView = inflater.inflate(R.layout.dialog_grid_item, parent, false);
+
+        int color = getColor(getColorResAt(position));
+
+        ImageView imageView = (ImageView) rowView.findViewById(R.id.icon);
+        if (color == selectedColor)
+            imageView.setImageResource(R.drawable.ic_checkmark_selected);
+        GradientDrawable gradientDrawable = (GradientDrawable) imageView.getBackground();
+
+        gradientDrawable.setColor(color);
+
+        return rowView;
+    }
+
+    @Override
+    public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
+        onColorSelected.onColorSelected(getColorResAt(position));
+    }
+
+    public interface OnColorSelected {
+        void onColorSelected(int color);
+    }
+}
